refactor: extract gateway control support helpers
